Make current SEPA connectivity controller independent
